Move thread functions to `base/thread.cpp/h`
