+++ CMakeLists.txt
@@ -2031,6 +2031,8 @@ set_src(BASE GLOB_RECURSE src/base
   str.h
   system.cpp
   system.h
+  thread.cpp
+  thread.h
   time.cpp
   time.h
   tl/threading.h
