+++ src/engine/client/backend_sdl.cpp
@@ -6,6 +6,7 @@
 
 #include <base/log.h>
 #include <base/math.h>
+#include <base/thread.h>
 #include <base/tl/threading.h>
 
 #include <engine/shared/config.h>
