+++ src/engine/shared/jobs.cpp
@@ -2,6 +2,8 @@
 /* If you are missing that file, acquire a complete release at teeworlds.com.                */
 #include "jobs.h"
 
+#include <base/thread.h>
+
 #include <algorithm>
 
 IJob::IJob() :
